A basic poc for the msp430g2553
GPIO pins can be configured for input/output, no pullup or pulldown
available just yet. The example which was supposed to blink a led doesn't
work because the button can't be configured correctly.

The main idea for this poc is to prove that we can do this with templates
and without using static methods like some other libraries do. This
results in ugly code with lots of '::' characters.

Led is a template with Pin type as parameter.
Pin is a template with the port as parameter.

For now this does result in an ugly Led instantiation:
Led<Pin<P1IN, P1OUT, P1DIR, P1IFG, P1IES, P1IE, P1SEL, P1REN, Gpio1Pins,
GPIO1_6> > led;