First working blinking led and MSP430G2553 configuration
Looks like the sample app didn't change at all except for the init (which
is necessary as far as the datasheet goes). That's a good thing since it's
one of the mail reasons I'm writing this library. Public api should change
as little as possible.

GpioTemplate has been renamed into GpioPort and a split between code with
and without interrupts is foreseen. Not all GPIO ports on the MSP430 range
have interrupt capability, so this is necessary if we want to have a clean
API.

Note: we can't call the init method from the GpioPort constructor, it
pulls in class-related code which we want to avoid. Why this is different
from the Button and Led classes isn't clear right now.