Added templates for gpio with and without interrupt + init
Looks like the sample app didn't change at all except for the init - which
is necessary as far as the datasheet goes.

While doing this I've added the full GPIO 1 & 2 for the MSP430G2553
device. This is useful anyway.

Note: we can't call the init method from the GpioPort constructor, it
pulls in class-related bits which we want to avoid. Why this is different
from the Button and Led classes isn't clear to me right now.