Different approach with peripheral pins
It's not as easy to find the right gpio pin when you need one, but it does
allow selecting pin modes much easier. Each pin can be configured in a
certain way, without restricting programs to a certain way of working.
Pins aren't really fixed, it's perfectly possible to create an input of a
pin that is configured for output. No idea why that would be used, but at
least it's possible.

This does have the drawback that there's an include file for each package
of a processor. This will most likely triple or more the amount of include
files and does make it slightly more difficult to migrate from one
processor to the other within the same family. Nevertheless, the focus for
this library is to make using the device simple. Migration should still be
simple, but it is perfectly acceptable to force users to find out which
pins they actually connect to themselves.