DB Tools Device: initial linout of this tool
The device (singular?) tool is a simple command-line tool that insert initial
data found in datasheets into the sjabloon430 database. Input files are created
by scripts that scrape a bunch of MSP430 (and CC430) datasheets and build up
lists of datasheets, devices and orderables.
These are then inserted into a database, which in turn is exported into files
containing SQL statements.

Rebase note: this tool has several iterations, starting at a very basic one
which was used to develop both the database structure further while writing
code for the pinout tool.
In the end - and at the time of writing this still isn't possible - this tool's
branch will be rebased onto the pinout tool in order to share the DB export.