DB: SQLite3 database to auto-generate code from datasheets' contents
The main idea of the tools to use this data is to generate the necessary header
files for each device, making it much easier to extend support to a large
number of different parts without risking mistakes from manual use.
Also, it's a good idea to gain some experience with C & C++ libraries.

Tables are self-explanatory, but still some explanations:
- datasheet contains SLAS-numbers and revision versions/dates
    This makes matching of pins easier, since pretty much only devices in the
    same datasheet share common pinouts.
- device contains data from a list of devices exported from the TI website and
  then modified a bit for easier parsing, exported again into a tab-separated
  file.
    This is actually a dumbed-down AND INCOMPLETE listing of features, but it's
    a first start aside from parsing datasheets, most likely at least partially
    manually.
- feature contains keys that are found in the TI export under the "Features"
  and "Special I/O" headers, amongst others.
    It should allow us to actually enable certain code only if a certain feature
    is present (e.g. ADC) etc. It has parameters because some features will
    require these without having to add a different Feature for each variation.
- device_feature is a rather simple link table between device & feature, also
  including parameter values.

Note that a previous version used many more technical ids (integers). After some
time I decided to use as little as possible because the idea is to export the
data in this database to SQL files. Files are re-generated during development in
order to fix issues - and the tools need to be tested. This may result in id
changes, thus a change on a large amount of lines which makes comparing/merging
much more difficult.