DB Data: output of first version of device tool, for development of pinout tool
This data - known to miss some data, acceptable - is required to write and test
the pinout tool.

It contains devices, packages and orderables of the currently known-to-me MSP
datasheets. This data was created from converted the pdfs to text and splicing
out that which I require. It is not guaranteed to be perfect though.