DB: pinset concept and links to orderable & signal
A link between devices and signals is important, combined with pin numbers.
This is required for generating pin-specific code, which is part of the latest
(experimental) code.

Pinset is the concept that fills this void. The idea is to link pin numbers
and signals. Pinset itself can then be linked to devices. But since pinsets
are also dependent on the package, it means there will be a pinset for each
package. Thus, it makes sense to use orderable to link to pinsets and make
that table the interlinking table.