cccurses: stdscr singleton, overloaded add() functions & raw int display
The overloaded add() functions create much cleaner code, in the future more
optimisations are possible as well.

Raw chtype display is logical since ncurses doesn't use char values, but
unsigned integers. It's also the only way to use the ACS_ constants directly,
they won't work when cast to char.

Regarding the standardScreen() function this appears to be the best solution
with a singleton.