DB Tools Pinout: read database files into in-memory database
Requires C++ 17 for std::filesystem etc.

The path to the database is hard-compiled into the binary, specified in the
CMakeLists.txt file. This is fine, it's a tool to manage the database for this
specific library with a certain structure. Anything that uses this code isn't
supposed to be used outside of this codebase

The import function requires a callback, which receives the filename processed
along with any errors. This makes it easy to share this in both command-line
applications and TUI/ncurses ones and any others.