DB Tools Pinout: separated database creation/import into separate source file
Creation & importing the database is best kept in at least one separate file.
It was fine while creating some mockups, but in the end readability greatly
improves by moving the database-related functions (both current and future)
into a separate file.

Much later, after creating the export functionality I deciced it would be nicer
to make even more distinction between database query functions and saving to
SQL text files. Therefore another file was added, which in this commit only
houses the import functionality.

The importDatabase() imports provided SQL files (with INSERT statements) into
the in-memory database. It uses a callback function to communicate some basic
information back into the main program. In this case all that does is simply
write filenames to the screen, along with any errors that may have been present
in that file.

This should be sufficient for the tool I'm currently envisioning; it's not a
user-facing utility.