DB Tools Pinout: db totals odd but efficient
New version of DB total counting to reduce complexity of the C++ code by adding
complexity in the query. Having a single row of counts and a single column for
each table makes it much easier to to retrieve them later.

This function is the first in which I create the SQLite prepared statement only
once. This is the better way, since SQLite keeps prepared statements internally
which in turn means multiple calls to this function would have turned in some
sort of a memory leak. For those afficionado's of other languages: both Java
and Rust wouldn't have been to fix this either. Of course I would be surprised
if a Java-based project like Hibernate would do an internal query-cache, but
it does make sense to do so actually.