DB Tools Pinout: [MOCKUP] search datasheet in centered window
Mockup for a datasheet search in a centered window by datasheet and model. Both
searches do a search for a datasheet. Model search is simply a convenience
method to search datasheets containing a certain model.
Pinouts are supposed to be edited/created for all devices of a datasheet, not a
single model only.