cccurses: Form buffer access & string trimming
Functions to access the buffer (single one, no support for user buffers yet)
contents. Currently the string value is trimmed from spaces, so it won't be
possible to add spaces in the beginning or at the end.

There are actually ncurses trim options as well if I recall correctly, so this
code may change later.