cccurses: extracted key input handling into separate code
This separated way of working makes it easier to customize handling without
losing the ability of using a predefined sensible default.
Form has become a template which statically uses the key handling. There is no
point in making this dynamically changed after construction. Templates are used
to simplify and more important increase efficiency.