cccurses: Form add field(s) + repost & Window clear became erase
Forms can be changed after creation, but they do require a unpost/post sequence
in order to work. In order to add fields they're added to a vector, which
replaces the older fixed-size array. This vector can also be dereferenced into
a FIELD** pointer, which makes copying completely useless.

Fields can now set their color as well, which is useful for alternating Field
colors (e.g. each row has a different color). This allows putting Fields right
below each other without a gap and still be distinguishable from the other
fields.

The change to Window is motivated because clear is basically an erase followed
by a refresh. The current erase() function does basically that, which may be
a bit unpredictable.