DB Tools Pinout: further mock-up of pin editing, dynamic signals
In this mock-up signal descriptions are retrieved from a map, i.e. no database
look-up but a single lookup to put all signals in memory.
This is suitable for production use, the number of signals isn't expected to be
truly massive.

This version also adds an empty field when no more empty field is available.
Note that the maximal number of signals is known (curretnly 9). This can be
used to reduce the amount of redrawing eventually.