cccurses: Form fields array requires duplication
Previous version of code used vector::data() to create the curses form, but
this vector can be modified later which doesn't sit well in certain cases.
Fun fact is that it seemed to work before, but extensive testing showed that
it was more or less an unlucky combination of circumstances.

Simply copying the array (on the heap, otherwise it'll be gone after function
return) fixes all issues. Only question is if this new'ed array is also deleted
after reposting. I'm guessing this array should be removed, but from the man
page of newform I'm expecting that freeform() should be called. Is this right?