DB Tools Pinout: keeping pins & signals in-memory + visualization
While there is still no actual usefulness to the program, this version at least
gives some indication of how the tool is supposed to operate.
Each pin/signal assignment is separated from the next one with a horizontal
line. In the future it would probably be better to print the packages more
frequently, this version only prints it once at the top of the list and once
on top of the editing pin.

Earlier I saw a window clear on each form repost, which was annoying. After
investigation this seems to be caused in the pinout code itself: the sub-window
of the form which was created always overlapped the first 10 lines of the main
window. This gave the perception that the entire window was cleared, but most
likely it were only the first 10 lines.