DB Tools Pinout: reordering packages feature
Having the possibility to reorder packages in the same order as they are
mentioned in the datasheet is required for sensible error-proof pinout editing.
The current version is implemented with regular ncurses wgetch() calls and case
switching. I'm not convinced that the current cccurses input handling will
result in considerably shorter code. It would however move some of the code
outside of this function, which may or may not confuse readers.