DB Tools Pinout: in Pinview, storing pins in a map
Storing pins in a vector makes sense, assuming the ordering of packages never
changes. The current version doesn't allow this, but it would make sense to
allow it. In fact, currently reordering is always prompted to the user. It may
be enough to allow it as a choice, which can then still be done after editing
some pins.