DB Tools Pinout: general SQL export to persist in-memory database
Exporting to SQL files is the best way to save a SQLite database for this
project. It's the reason I chose an in-memory database, because there is no
non-binary file format in SQLite. In fact, it surprises me that there is no
export to SQL function directly in SQLite itself.

The provided function exports a single prepared statement. The reason I chose
this above exporting a single table is that I want to create SQL files for each
datasheet id (e.g. SLAS555) instead of a single big export. This makes it
easier both to add and to check which pinouts have been already created.