DB Tools Pinout: (db::)Package struct for passing packages
Passing packages as single concatenated strings (drawing & # of pins) is fine,
until the point where we actually need to save pinsets. These are linked to the
packages, but the primary key for these is drawing & pins. Separating it later
is much harder than passing it as a struct and concatenating at places where it
is currently in use in a concatenated fashion.

It might be useful to allow Package to be used as the key for an unordered_map.
As far as I know that isn't really difficult, but it is currently not
implemented as such.