DB: using signal_group as lookup table for feature table
The previous solution where signal was linked directly to the feature table
didn't work. This was most likely because the values in the referenced column
weren't unique, possibly worsened by the lack of constraints and indices.

Note that I chose to not reorder the files (i.e. feature should come after
signal), but chose to alter the table in the signal script. This may be changed
eventually.

As far as that change goes: SQLite doesn't understand what
ALTER TABLE ... ALTER COLUMN means and requires drop & add. Another reason to
choose reordering after all.