DB Tools Pinout: signal description field dynamic and better sizing
For ncurses, a dynamic field means that the data size is different from the
field's size.

This tool benefits from also sizing the field based on a good limit of 80
columns on the screen. Simply put, it's size to end in column 79, unless the
minimal size isn't reached. Then it will reach further, but I don't expect this
to happen very often, if ever.