DB: signalset concept, link between pinset & signal
Previously the idea was to link signals directly to a pinset. This isn't
necessarily impossible, but it creates a few issues.

Assuming that a pinset is linked to orderables, every package (or at least pin
count) requires a separate pinset i.e. 20-pin & 28-pin packages cannot share
the same pinset. This in turn means that each pinset maps to signals separately
but this duplicates content. It means that things may diverge, either from tool
bugs or manual mistakes. Normalisation into signalset fixed this, with each
pinset linking to the signalsets involved.

Also filtering/parenting on signalset level for different models or packages
is required and needs to be reconstructed when editing.

On top of that it is much easier to map this to a UI that is inspired by the
datasheet layouts, rather than deduplicating the data in the database.

Note that some experience with writing the pinout database tool has driven this
requirement. Concept of hindsight basically, rebase-driven development ;)