DB Tools Device: temporary tool broken after rebasing on latest database branch
Ok, so the device tool that I initially wrote was a quick 'n dirty experiment
to test some stuff and get my bearings. It was sufficient to develop another
tool - the pinout tool - which required this data.

The current version of the tool not only doesn't work anymore, but it's rather
ugly, difficult to maintain and as such requires extensive rework.
On top of that, it didn't create any SQL exports yet, these were done manually
from a file-based database that it simply filled.

In short, commits after this commit will remedy this.

Earlier commits will not be reworded anymore, even though they were far from
my usual quality.