DB: signalset datasheet index for ordering in tools
Datasheet index is a purely view-driven property, it doesn't really change the
data/links directly. In fact I would prefer to be able to follow a certain
ordering which can be deduced from existing data in the datasheets.

But for editing in tools the most important part is that it is possible to put
them next to a datasheet and edit them with some confidence.