DB Tools Pinout: DB import status now scrolls if need be, clearer ERROR as well
By scrolling the status message if need be, there won't be too many lines
displayed for the screen. Of course this does mean that previous errors are not
visible. If there are many more files in the future, it might be best to
display either only error-ing files or display all files on a single line
except the ones with errors.

Another option might be to display files that contain errors first by moving
the cursor back, although that might be a mess if the window has actually
scrolled further.