cccurses: Window::print() function taking vararg
I thought I could have passed on these varargs with std::forward, but this
doesn't seem to be true. va_list/va_args/va_end is the correct way to do this,
at least in C. std::forward did also work though, so I expect this to be a
possible solution as well. However, I had an issue with the first parameter not
getting passed on correctly.

Both solutions do however emit a compiler warning that needs to be silenced
with a pragma. Disabling it in the CMakeLists.txt isn't an option, since the
C++ wrapper for ncurses is assumed to be used header-only.