DB Tools Pinout: db counts display for smaller terminals & supported percentage
The DB count would fail printing in the top window for terminals 130 columns or
less wide, which is not sufficient. In fact, I really want this application to
support terminals of 80 columns wide as well.

Since the DB count had to be modified anyway, I chose to add a second db total
to count the percentage of datasheets/devices/orderables. While technically not
part of the MVP (as it is purely informational), it doesn't add that much extra
development anyway.

Note that I removed the package count. While it is informational, it doesn't
really add much and it is difficult to fit in a tight terminal.