DB Tools Pinout: export sql with update statements
Not all exports can be created as insert statements, sometimes updates will be
necessary. This generic update statement generator adds this functionality.

The first columns in the statement are required to be the conditions, i.e. the
columns that will appear in the WHERE part of the update. The amount of columns
is assumed to be 1 unless specified.