DB Tools Pinout: exporting orderables both general & for a specific datasheet
Orderables are the links between devices & datasheets on one hand and signals
(through signalset and pinset) on the other hand. But they are also an integral
part of a model, extracted from a datasheet. Therefore I prefer to keep them on
top of the SQL export file.
Pinsets are not yet in the database at that point and cannot be linked to until
at the end.

Ordering by pinset_id and device.model is important to group related statements
together. Sharing pinsets means the devices have the same pin numbers and many
orderables sharing the same packages (and config set - not yet implemented)
belong together.

On orderables (& devices) export:

Originally the idea was to concentrate as much data into datasheet-specific SQL
files. This change moves the orderable data for a specific datasheet into the
datasheet-specific one. Since orderable data is updated to link pinout data,
this makes sense.

The question now arises if there's a benefit - aside from being cleaner - to
separate all data for a single datasheet. It does make sense to do so, but adds
quite a bit of code which isn't strictly necessary.
A quick check makes me consider following:

- datasheet records: if there is an update the datasheets - even though MSP430
    isn't really supported, they still might get updatess. This adds revisions
    etc. Currently the only way this data is exported into SQL files is by
    making the modifications inside a file-based database and then exporting it
    with an external tool. If the tool that updates the database records does
    this too, it would be fine having the record in a separate datasheet-related
    file.

- device records: these are less likely to be modified but they are also added
    and updated partially manually. The same considerations as datasheet apply
    here with an extra consideration: when a device is added to one of the
    existing datasheet this will turn into devices in two separate files.

In fact, both of these also apply to orderables since they're directly linked
to devices. It might make sense to keep device & orderables together while
keeping the datasheet data in a separate file for now.