DB Data: SLAS735, for testing configsets/parenting
Added a few pins that don't require configsets right now. This functionality is
supposed to work but after changes broke several times.

Note: this pin assignments wrongly uses signals:
 VREF-/VEREF-
 VREF+/VEREF+

They need to be replaced with VREF-, VEREF-, VREF+ and VEREF+.
The signals export does contain these, with a warning as description not to use
them. I thought this was best as they may occur in other datasheets resembling
a single signal.