DB Data: SLAS735, for testing configsets/parenting
Added a few pins that don't require configsets right now. This functionality is
supposed to work but after changes broke several times.

Note: this specific file no longer imports at this point because a few of the
mentioned signals no longer exist in the rebased 24_signals.sql file:
 VREF-/VEREF-
 VREF+/VEREF+
If it is important to read in this file, remove these lines but they're kept
since they're fixed later on anyway.