DB Tools Pinout: read data from DB and display it
I've tried keeping DB-scoped objects out of the view as much as possible. The
view itself isn't directly linked to the DB-data, which is conceived to make
generation of code/headers possible.

This is still far from MVP and may possibly change, but I've done my best to
keep options open for later.