DB Tools Pinout: saving pinset functionality (NO scrolling, no configsets)
Previously pinsets weren't really saved. There had been a temporary save, which
had then become defunct. (this may have been rebased out) Originally I started
with a pinset/signalset created by the application, without having any reading
functionality. This didn't appear to be the best method, since the issues arose
mostly during reading (and still will, config sets have yet to be implemented).

Pinsets are created and linked to Orderable before editing, one pinset for each
package. After converting the view back to database objects, the same function
is called again to associate signalsets with pinsets.

Pinsets are saved one by one. This makes it simpler and in the end the amount
of pinsets is rather small.

Note that this is the first version and - hindsight altert - things will change
later to accomodate configsets and because the logic could be simpler.Â