DB Data: SLAS735 manual pinset parents for configset development
In my language we say: "a donkey won't hit his head on the same stone twice".
Previously I chose to develop saving/exporting before importing, thinking that
it would be easier to export the data first and then develop reading it.
It turned out that several pitfals were present on re-creating the entire view
from the database and to prevent this I'm starting from the import side this
time.

So, pinsets are parented and assigned to their orderables.
That said, this export was already created a bit later from a partially working
version.

Note: I removed MSP430G2113IN20 linked to MSP430G2153 because it's functionally
incorrect. It'll be restored later, when there is actually a device for it.

Also formatted column widths, this older style is no longer saved by the code.