UI fix scrolling more or less bug TODO REBASE? RE-ORDER?
Bugfix for a small ui scrolling bug:
when scrolling all the way down, pressing enter to edit (or rather: insert a
new signalset) and then cancelling this edit again (by pressing enter). Since
there are no signals or pins, no data is added. When the user in this case
presses enter to edit again, all data is scrolled up without erasing other data
since it is only necessary to redraw when not editing. This edge case did not
fit this assumption.

Another fix is - when editing the last set again - that the amount scrolled was
only the required minimum of 2 rows. But again, when un-editing and then scrol-
ling down the window kindof "jumps". It is really minor, but the fix is as well.