DB Data: SLASEE5, MSP430FR2422
Value line part, one of the cheapest and yet 7.5k FRAM & 2k SRAM. Not part of
the LaunchPad or other EVM types, just a simple MCU which might be one of the
best choices for many projects.
Pretty much the smaller brother to the MSP430FR2433, which does have a Launch-
Pad; but comes in less hand-solderable QFN & BGA packages only.