DB Tools Pinout: configset window fallback for small terminals (& LLDB)
I already knew that the application wouldn't work when the set window outgrows
the terminal height for some time, but until configsets were really improving
implemented it wasn't a priority. Since it also broke debugging it became one.

The code has become rather complicated with several ifs added, there might be
a cleaner option but the functionality is there.
The fallbacks are quite numorous:
* leaving out the default set if all others can be rendered
* leaving out packages of all sets (doesn't really work with package-only set)
* leaving out headers
* and if all else fails: add a single line per configset, without models/pkgs