DB Data: SLAS735 manually removed duplicate pinset between N20 & PW20 REORDERED
The idea is to have the least amount of duplicate data and when pinsets are
shared in the datasheet it's nice to do this in the DB as well. The application
doesn't do this itself (could be a nice addition, but there are many pitfalls).
Manually though the decision is a lot easier and can be made after doing all of
the inserting/editing.

Should someone read this commit and wonder how I did this without errors, it's
actually rather simple. All SQL in the SQL files is executed and a simple way
to update this is to add manually-generated SQL statements at the end (before
the COMMIT TRANSACTION ofcourse).
More specifically for this file (check the previous version of the file!):
1. assign all orderables the lower id for N20&PW20
    UPDATE orderable SET pinset_id=5 here pinset_id=6;   -- PW20=N20
    UPDATE orderable SET pinset_id=9 where pinset_id=10; -- parent!
2. execute the pinout tool! open this sheet, save it again. This is where
    it all happens: the ids are updated and the export NO LONGER contains the
    other pinsets because they're no longer linked to - that is why the parent
    is of course also updated. Since they're no longer in the export, they no
    longer exist.
3. remove gaps in the ids (note that there were no other ids in the db)
    UPDATE pinset SET id=id-1 WHERE id > 10;
    UPDATE pinset SET id=id-1 WHERE id > 6;
It is possible that an operation like this results in an error, but in
this case I believe it worked. IF it didn't, I simply updated the id first
with 'SET id=id+100' and then did 'SET id=id-101'.

Note that this works because pinset.id is cascaded for update on all foreign
keys. The same is true for signalset.id.

Also note that this ALSO works if there is more data in the database, since
only the data FOR THIS DATASHEET will be exported. However, it is of the
utmost care that there are no duplicate ids!

This trick can also be used if ids *are duplicated on different branches or
in different files.