DB Tools: database creation now enables foreign keys
Ok, apparently the foreign keys aren't checked without this PRAGMA. This isn't
what I expected to happen especially since these worked in the GUI editor that
I use.

Of course this is a change that was moved here using rebasing, so I'll simply
explain what happened.

While developing the pinout tool there are a few datasheets for which I had
entered the pinout data to test all functionality. One of them (and this will
happen much more often for other sheets) has a duplicate pinset, also witnessed
in the datasheet.
The tool itself doesn't support this, it creates a separate pinset for each
package option that exists and again that many for each configset that is
created. (configsets are the equivalent for datasheets mentioning things like
'MSP430XYZ devices only').

The easiest way that I can think of without coding it into the tool - which I
still may, but across multiple datasheets - is by manualling inserting update
statements (on the pinset table). These are then cascaded into other tables,
but this is exactly that which isn't working without the PRAGMA statement.