DB: pinset id view to simplify queries and include all parents
During development of the pinout tool it became clear that certain queries were
rather large and still didn't meet their requirements of matching all pinset
parents and further. In the end I chose to add a simple view to improve the
queries and it really made the biggest of differences.