DB Tools Pinout: configset indicates amount of orderables
For someone who doesn't know what the number mean it might not be very useful,
but having the orderables selected next to the configset is rather useful:
both at creating and editing configsets it adds an extra verification.
I've noticed this myself when updating the data, which added orderables to some
datasheets. After editing the sets it was clear each time that extra orderables
were added.