DB Tools Pinout: removed pin count from pinset
After writing the pinout tool it became clear that calculating the pin count
isn't that simple anymore, especially configsets & parenting.
In fact it's easier to count/list it using queries should it be necessary.

Originally the idea was that it could be used to query for a certain pinset to
make it easy to find it and link to it. It may still be useful for later ideas,
but then it should be correct anyway.