DB Tools Pinout: remove gaps in signalset ids TODO: does not yet remove signalsets; shouldn't it re-use ids? better options? do it in the view code?
In complex datasheets it is possible to do some editing, save and pick up
further editing later. When this is used to add configsets at a later stage or
moving signals between configsets a new signalset will be added. The old one
may be left empty. This empty signalset is then removed from the SQL exports,
leaving a gap in the ids.

TODO: this feature (at least when it's corrected to remove unlinked signalsets)
will result in more difficult merges? Updating it manually is a safer option?