DB Tools Device: datasheet update & insert without export
The data is read from the TSV file, converted into an object and then inserted
into the database. This is a bit more than I did before, where I would create
the insert statement and simply execute it (albeit without using parameters).
If there is any performance penalty, it won't matter for this usage.