DB Tools Device: orderables reading & export
Orderables are now also added, meaning the application is fully functional
again. MSL levels have been extended to 4 since I've seen several datasheets
which use this.

I have kept one database function of the old code: device matching. It's a good
one in that it actually does its job well, but it can't really be rewritten
nicely since it has to build up the query dynamically for each orderable.
The one possible improvement is adding the datasheet as a parameter instead of
concatenating it in the string. I'll consider this acceptable for now.
Should I change my mind I can always rewrite it, but with no real benefits.