DB Tools Device: rewritten device matching and fixed issues with missing MSL
Of course I was going to rewrite that matching function as well, what else did
you expect? Rewriting a perfectly working function into something a bit cleaner
is the perfect use of my time in this thing I used to call my life.

Also fixed an issue where some records would not contain a MSL value (Call TI)
and miss a column. Since MSL, temperature etc aren't really required, this is
a good place for a fallback. If there is no MSL, it is set to level 5 and the
next token is assumed to be the temperature.