DB Data: export of all data from second version of device tool and data merge
While developing the pinout & device tools this happened on a separate branch.
The first version of the device tool was used to generate initial data for
developing the pinout tool. The pinout tool also developed the final format and
export code for the SQL files. This was then incorporated in the second version
of the device tool.

Now the circle is round, with the data from that updated device tool merged
into the tools-pinout branch.

Note that:
* 21_device.sql no longer exists; this data is in the 90_ files per datasheet
* 26_orderable.sql no longer exists; same reason as above
* 23_package RTD added
* headers in all files are now exported by the tools