DB Data: SLAS735 assigned pinset to 'new' orderable, BSL fix & format update
There was one device in this datasheet that wasn't present in the earlier
versions of the database. Due to a data error I removed it.
After merging the latest data from the 'device' tool, this one has now been
correctly added and thus also needed to be assigned the correct pinset.

This also updates the data to the latest formatting & column definitions: the
group_idx columns has been added to pinset and spacing is corrected as well.

The BSLRX signal was also missing, so this has been corrected.

When diffing this commit it's easiest to diff without whitespaces.