DB Data: SLAS735 assigned pinset to 'new' orderable, pinsets with group_idx TODO REVIEW fixed BSL
Another device was added to link the correct orderable (which was removed
temporarily to prevent bad data). This adds this again after merging the latest
DB from the pinout tool.

This also updates the pinset data to include the group_idx, which contains the
configset index that is used in the pinout tool.

BSL signals, one disappeared...  TODO REVIEW!