DB Data: SLAS272 added pinsets for RTD package & group_idx in pinset
Previously missing from the device/orderable SQL data, these orderables were
added with the latest import of the updated device tool.

Pinset ids have been removed to share them again between all packages, as it is
in the datasheet. Nonetheless this was a good test for the pinout tool which at
this point is working as required.