DB Tools Pinout: BUGFIX configset parenting with pinset id 0 orderable links
Sometimes configsets have different packages, either manually specified or
because the models in that configset only exist for a subset of packages.
There was a small bugfix in the code which was supposed to fix this, but that
fix in itself resulted in the odblPinsetLinks map getting updated with 0 ids.

The orderable linking code assumes that the pinsets already exist, which is not
true for a pinset with id 0; this results in failures in the query, meaning
nothing is added to the export files.

There is another important fix in the parent set identification. Before it
would assume the parent set only needs to assume a single shred orderable.
There are cases where this is not true, especially when filtering on packages.